Added e2e tests for dependency isolation (#919)
* Added e2e tests for dependency isolation
