+++ tests/endtoend/dependency_isolation_functions/report_dependencies/function.json
